2022/day21/day21_test.go
package day21

import (
	"fmt"
	"testing"
)

func TestP1(t *testing.T) {
	got := part1("sample.txt")
	fmt.Printf("Got: %v\n", got)
	want := 152
	if got != want {
		t.Errorf("Got %v but want %v", got, want)
	}
}

func TestP1_Actual(t *testing.T) {
	got := part1("input.txt")
	fmt.Printf("Got: %v\n", got)
	want := 145167969204648
	if got != want {
		t.Errorf("Got %v but want %v", got, want)
	}
}

func TestP2(t *testing.T) {
	got := part2("sample.txt")
	fmt.Printf("Got: %v\n", got)
	want := 301
	if got != want {
		t.Errorf("Got %v but want %v", got, want)
	}
}

func TestP2_Actual(t *testing.T) {
	// debug.SetMaxStack(1000000000000000)
	// 7560831729513 - too high
	// 7560831729513
	got := part2("input.txt")
	fmt.Printf("Got: %v\n", got)
	want := 3330805295850
	if got != want {
		t.Errorf("Got %v but want %v", got, want)
	}
}
